backend/requirements.txt
click==8.1.3
colorama==0.4.5
Flask==2.2.2
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpickle==2.2.0
MarkupSafe==2.1.1
marshmallow==3.17.1
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
python-decouple==3.6
python-dotenv==0.20.0
six==1.16.0
Werkzeug==2.2.2
psycopg2-binary==2.9.5
